Create Lab Repot
Symbol table is an important data structure created and maintained by compilers in order to
store information about the occurrence of various identifiers such as variable names, function
names, objects, classes, interfaces, etc. Symbol table is used by both the analysis and the
synthesis parts of a compiler. Symbol table can be implemented in one of the following ways:
• Linear (sorted or unsorted) list
• Binary Search Tree
• Hash table
• And other ways.
In this lab session, you are required to analyse the various implementations. You need to write
code for at least two ways of implementation. Test your code with different test cases. Submit
a report of your analysis and executable code by the end of the session.